pyproject.toml
[project]
name = "ai-engineer-projects"
version = "0.1.0"
description = "Projeto desenvolvido para a certificação de AI Engineer na Indicium."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "crewai>=1.4.1",
    "duckdb>=1.4.1",
    "groq>=0.34.0",
    "ipykernel>=7.0.1",
    "langchain>=1.0.5",
    "langfuse>=3.9.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openinference-instrumentation-crewai>=0.1.16",
    "openinference-instrumentation-litellm>=0.1.28",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "seaborn>=0.13.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["data", "docs", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4"]
ignore = ["E501"]  # exemplo: ignora erro de linha longa
fixable = ["ALL"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.0",
]
